Restructure gast to code into 4 files
