+++ debug.py
@@ -1,8 +1,4 @@
-#import gast_to_code.gast_to_code_main
 import code_to_gast.js_main as js_main
 import code_to_gast.py_main as py_main
-#import code_to_gast.routers.py_router
-#import code_to_gast.helpers.py_helpers
-#print(py_main.py_to_gast("x = True"))
 
 print(py_main.py_to_gast("x = True"))
